refactor: Centralize config args
